backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.21
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.8.0
